every time-related int is now long long (except diff)

+++ mc.cpp
@@ -41,7 +41,7 @@ public:
         return lattice[((x + Nx)%Nx)*Ny + (y+Ny)%Ny];
     }
 
-    void metropolis(int Nsteps, double T, Measurer *m);
+    void metropolis(long long Nsteps, double T, Measurer *m);
 
     // Lattice parameters
     int Nx, Ny, Nsites;
@@ -172,12 +172,12 @@ void IsingMCWorker::print_J_lattice() {
     }
 }
 
-void IsingMCWorker::metropolis(int time, double T, Measurer *m = nullptr)
+void IsingMCWorker::metropolis(long long time, double T, Measurer *m = nullptr)
 {
     uniform_int_distribution<int> distx(0, Nx-1);
     uniform_int_distribution<int> disty(0, Ny-1);
     uniform_real_distribution<double> distmc(0, 1);
-	for (long int i = 0; i < time; i++)
+	for (long long i = 0; i < time; i++)
 	{
         for (long int j = 0; j < Nsites; j++) {
             int x = distx(engine);
@@ -278,7 +278,7 @@ long long Measurer::avg_helper(vector<long long> times, long long maxtime)
     if (times.size() == 0)
         return maxtime;
     long long sum = 0;
-    int curt = 0;
+    long long curt = 0;
     int flag = +1;
     for (int i = 0; i < times.size(); i++) {
         sum += (times[i] - curt)*flag;
@@ -417,7 +417,7 @@ int main(int argc, char* argv[])
     }
 
     // Thermalization
-    long Ttherm;
+    long long Ttherm;
     cmd("--therm", 0) >> Ttherm;
     if (Ttherm > 0) {
         cerr << "Thermalizing... ";
@@ -429,7 +429,7 @@ int main(int argc, char* argv[])
         mc.print_system();
     }
     // Measuring
-    long Tmc;
+    long long Tmc;
     if (!(cmd("--time") >> Tmc)) {
         cerr << "Please provide Monte-Carlo time!" << endl;
         print_usage(cmd);
